+++ assets/javascripts/kitten/components/molecules/cards/project-card/index.js
@@ -26,6 +26,7 @@ export const ProjectCard = ({
         className,
         `k-ProjectCard--${status}`,
       )}
+      href={href}
       {...props}
     >
       <div className="k-ProjectCard__image">