+++ assets/javascripts/kitten/components/molecules/cards/project-card/stories.js
@@ -83,6 +83,14 @@ export const Default = ({ progressProps, ...args }) => (
 Default.args = args
 Default.argTypes = argTypes
 
+export const AsLink = args => <Default {...args} />
+
+AsLink.args = {
+  ...args,
+  href: '//www.kisskissbankbank.com',
+}
+AsLink.argTypes = argTypes
+
 export const WithVideo = args => <Default {...args} />
 
 WithVideo.args = {