+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `ProjectCard`: Fix link props.
+
 ## [6.12.0] - 2021-12-28
 
 Features: